# internal/api/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: Aether Notebooks API — collaborative SQL/data notebook platform
  title: Aether API
  version: 1.0.0
paths:
...
  /attachments/{id}:
    delete:
      description: Delete a file attachment by ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete attachment
      tags:
      - attachments
    get:
      description: Download a file attachment by ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get attachment
      tags:
      - attachments
...
  /notebooks/{notebook_id}/attachments:
    get:
      description: List all file attachments for a notebook
      parameters:
      - description: Notebook ID
        in: path
        name: notebook_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List attachments
      tags:
      - attachments
    post:
      consumes:
      - multipart/form-data
      description: Upload a file attachment to a notebook
      parameters:
      - description: Notebook ID
        in: path
        name: notebook_id
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload attachment
      tags:
      - attachments
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"